Add a new script to fix some annoyances on reboot
Specifically the script will restart all of the apps that are bound to all desktops.
macOS has a very annoying bug where apps bound to all desktops are incorrectly bound when restarted after a reboot.

The windows of the apps are only visible on the first desktop.
When entering exposé on another desktop there is some space for these windows, but the windows are still completely invisible.
The invisible windows are highlighted on hover, but selecting them does not solve the issue: the windows stay invisible.

The only way to fix this is relaunching the apps (presumably on the first desktop, but might work from another desktop).